+++ apps/br_ipti_tag_app/lib/app/features/student/domain/usecases/create_student_usecase.dart
@@ -3,6 +3,7 @@ import 'package:tag_sdk/src/core/defaults/usecase.dart';
 import 'package:tag_sdk/src/core/failures/failures.dart';
 import 'package:tag_sdk/src/features/student/domain/entities/student.dart';
 import 'package:tag_sdk/src/features/student/domain/repositories/student_repositories.dart';
+import 'package:tag_sdk/tag_sdk.dart';
 
 class CreateStudentsUsecase
     implements Usecase<StudentIdentification, CreateStudentParams> {
@@ -18,10 +19,4 @@ class CreateStudentsUsecase
   }
 }
 
-class CreateStudentParams {
-  const CreateStudentParams({
-    required this.student,
-  });
 
-  final StudentIdentification student;
-}
