+++ apps/br_ipti_tag_app/lib/app/features/teacher/teacher_module.dart
@@ -3,6 +3,7 @@ import 'package:br_ipti_tag_app/app/features/teacher/presentation/list/bloc/teac
 import 'package:br_ipti_tag_app/app/features/teacher/presentation/list/pages/teacher_list_page.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:modular_bloc_bind/modular_bloc_bind.dart';
+import 'package:tag_network/tag_network.dart';
 import 'package:tag_sdk/tag_sdk.dart';
 
 import 'domain/usecases/list_teachers_usecase.dart';
@@ -18,22 +19,24 @@ class TeacherModule extends Module {
   final List<Bind> binds = [
     // datasources
     Bind.factory((i) => TeacherRemoteDataSource(
-          i.get(),
-        )),
+      i.get<RouterAPI>(),
+    )),
 
     // repositories
     Bind.factory((i) => InstructorRepositoryImpl(
-          i.get(),
-        )),
+      i.get<TeacherRemoteDataSource>(),
+    )),
 
     // usecases
-    Bind.factory((i) => ListInstructorsUsecase(
-          i.get(),
-        )),
+    Bind.factory(
+      (i) => ListInstructorsUsecase(
+        i.get<InstructorRepositoryImpl>(),
+      ),
+    ),
 
     // list
     BlocBind.factory((i) => TeacherListBloc(
-          i.get(),
+          i.get<ListInstructorsUsecase>(),
         )),
   ];
