+++ packages/tag_sdk/lib/src/features/school/school_module.dart
@@ -5,7 +5,6 @@ import 'data/repositories/school_repository_impl.dart';
 
 class SchoolModule extends Module {
   @override
-  @override
   List<Bind<Object>> get binds => [
         // Datasources
         Bind.singleton((i) => SchoolRemoteDataSource(
@@ -14,16 +13,6 @@ class SchoolModule extends Module {
 
         // Repositories
         Bind.singleton((i) => SchoolRepositoryImpl(
-              i.get(),
-            )),
-
-        // UseCases
-        Bind.singleton(
-          (i) => EditSchoolUsecase(
-            i.get<SchoolRepositoryImpl>(),
-          ),
-        ),
-        Bind.singleton((i) => ShowSchoolUsecase(
               i.get(),
             )),
       ];
