+++ packages/tag_sdk/lib/src/features/student/domain/entities/create_student_params.dart
@@ -1,0 +1,9 @@
+import '../../../../../tag_sdk.dart';
+
+class CreateStudentParams {
+  const CreateStudentParams({
+    required this.student,
+  });
+
+  final StudentIdentification student;
+}
