+++ packages/tag_sdk/lib/src/features/teacher/teacher_module.dart
@@ -1,10 +1,7 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:tag_network/tag_network.dart';
-import 'package:tag_sdk/src/features/teacher/domain/repositories/instructor_repository.dart';
-
 import 'data/datasources/remote/teacher_remote_datasource.dart';
 import 'data/repositories/teacher_repository_impl.dart';
-import 'domain/usecases/list_teachers_usecase.dart';
 
 class TeacherSDKModule extends Module {
   @override
@@ -21,14 +18,6 @@ class TeacherSDKModule extends Module {
     Bind.factory(
       (i) => InstructorRepositoryImpl(
         i.get<TeacherRemoteDataSource>(),
-      ),
-      export: true,
-    ),
-
-    // usecases
-    Bind.factory(
-      (i) => ListInstructorsUsecase(
-        i.get<InstructorRepository>(),
       ),
       export: true,
     ),
